- Added some docs to the database code - Extracted methods for running statements to SqlDatabaseUtils.kt
